+++ tests/test_redis.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python
 
 from basic_store import BasicStore, TTLStore
+from conftest import ExtendedKeyspaceTests
+from simplekv.contrib import ExtendedKeyspaceMixin
 
 import pytest
 redis = pytest.importorskip('redis')
@@ -22,4 +24,23 @@ class TestRedisStore(TTLStore, BasicStore):
 
         r.flushdb()
         yield RedisStore(r)
+        r.flushdb()
+
+
+class TestExtendedKeyspaceDictStore(TestRedisStore, ExtendedKeyspaceTests):
+    @pytest.fixture
+    def store(self):
+        from simplekv.memory.redisstore import RedisStore
+
+        class ExtendedKeyspaceStore(ExtendedKeyspaceMixin, RedisStore):
+            pass
+        r = StrictRedis()
+
+        try:
+            r.get('anything')
+        except ConnectionError:
+            pytest.skip('Could not connect to redis server')
+
+        r.flushdb()
+        yield ExtendedKeyspaceStore(r)
         r.flushdb()